refactor: 패키지 표현 변경
