+++ src/main/java/in/koreatech/koin/domain/auth/JwtProvider.java
@@ -10,7 +10,7 @@ import javax.crypto.SecretKey;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import in.koreatech.koin.domain.user.domain.User;
+import in.koreatech.koin.domain.user.model.User;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
