+++ src/main/java/in/koreatech/koin/domain/shop/model/MenuCategory.java
@@ -1,10 +1,9 @@
-package in.koreatech.koin.domain.shop.domain;
+package in.koreatech.koin.domain.shop.model;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import in.koreatech.koin.global.common.BaseEntity;
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -21,11 +20,9 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @Entity
-@Table(name = "shop_menus")
+@Table(name = "shop_menu_categories")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Menu extends BaseEntity {
-
-    private static final int SINGLE_OPTION_COUNT = 1;
+public final class MenuCategory extends BaseEntity {
 
     @Id
     @Column(name = "id", nullable = false)
@@ -41,44 +38,16 @@ public class Menu extends BaseEntity {
     @Column(name = "name", nullable = false)
     private String name;
 
-    @Size(max = 255)
-    @Column(name = "description")
-    private String description;
-
-    @NotNull
-    @Column(name = "is_hidden", nullable = false)
-    private Boolean isHidden = false;
-
     @NotNull
     @Column(name = "is_deleted", nullable = false)
     private Boolean isDeleted = false;
 
-    @OneToMany(mappedBy = "menu", orphanRemoval = true, cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "menuCategory")
     private List<MenuCategoryMap> menuCategoryMaps = new ArrayList<>();
 
-    @OneToMany(mappedBy = "menu", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<MenuOption> menuOptions = new ArrayList<>();
-
-    @OneToMany(mappedBy = "menu", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<MenuImage> menuImages = new ArrayList<>();
-
     @Builder
-    private Menu(Long shopId, String name, String description) {
+    private MenuCategory(Long shopId, String name) {
         this.shopId = shopId;
         this.name = name;
-        this.description = description;
-    }
-
-    public boolean hasMultipleOption() {
-        return menuOptions.size() > SINGLE_OPTION_COUNT;
-    }
-
-    @Override
-    public String toString() {
-        return "Menu{" +
-            "id=" + id +
-            ", shopId=" + shopId +
-            ", name='" + name + '\'' +
-            '}';
     }
 }
