+++ src/main/java/in/koreatech/koin/domain/shop/model/MenuImage.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.shop.domain;
+package in.koreatech.koin.domain.shop.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
