+++ src/main/java/in/koreatech/koin/domain/track/repository/TechStackRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.track.domain.TechStack;
+import in.koreatech.koin.domain.track.model.TechStack;
 
 public interface TechStackRepository extends Repository<TechStack, Long> {
