+++ src/main/java/in/koreatech/koin/domain/user/service/UserService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin.domain.auth.JwtProvider;
-import in.koreatech.koin.domain.user.domain.User;
-import in.koreatech.koin.domain.user.domain.UserToken;
+import in.koreatech.koin.domain.user.model.User;
+import in.koreatech.koin.domain.user.model.UserToken;
 import in.koreatech.koin.domain.user.dto.UserLoginRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginResponse;
 import in.koreatech.koin.domain.user.repository.UserRepository;
