+++ src/test/java/in/koreatech/koin/acceptance/UserApiTest.java
@@ -8,9 +8,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
 import in.koreatech.koin.AcceptanceTest;
-import in.koreatech.koin.domain.user.domain.User;
-import in.koreatech.koin.domain.user.domain.UserToken;
-import in.koreatech.koin.domain.user.domain.UserType;
+import in.koreatech.koin.domain.user.model.User;
+import in.koreatech.koin.domain.user.model.UserToken;
+import in.koreatech.koin.domain.user.model.UserType;
 import in.koreatech.koin.domain.user.repository.UserRepository;
 import in.koreatech.koin.domain.user.repository.UserTokenRepository;
 import io.restassured.RestAssured;
